Preparation for ARM code
The fam_atomic library currently works only on x86, and the simulated
ioctls invoke the simulated atomics that are prefixed with x86 (x86_xadd,
ect...). In preperation for also providing support for ARM, this patch
modifies the naming convention so that we call general functions (xadd
instead of x86_xadd, ect...) and add an ifdef __x86_64__ which #define
the general functions to the x86 version of the functions. Thus, when the
arm code gets added, we can add an ifdef __aarch64__, and under the ifdef,
define the general functions to call the arm version of the functions.

Signed-off-by: Jason Low <jason.low2@hpe.com>