Modify method to raise segmentation fault on unregistered atomic usage
There has been a bug report stating that a script forking an
executable using an unregistered fam atomic caused the script
to segfault instead of the executable itself, making it
difficult to debug the issue.

This patch addresses it by modifying the code so that we
raise the segmentation fault using the raise() function.

Signed-off-by: Jason Low <jason.low2@hpe.com>