tests: Make single node test run for specified duration
The fam_atomic_single_node_test tests the atomics by calling
some fixed number of atomic operations. The problem is that
the duration it takes to complete the test can vary
significantly between systems.

This modifies the test to make the test duration more
deterministic by having it run for some specified duration
in seconds instead of performing a fixed number of operations.

Signed-off-by: Jason Low <jason.low2@hpe.com>