Specify extern "C" in fam_atomic.h when included by C++ applications
The libfam-atomic library is compiled in C, so when the fam atomic
interfaces are used by C++ applications, the compiler will need to
know that these functions are compiled in C. This can be done by
wrapping the include fam_atomic.h in an extern "C".

We'll handle specifying extern "C" in the header file so that
C++ applications don't need to worry about these details.

Signed-off-by: Jason Low <jason.low2@hpe.com>