Allow fam registration to work across threads
Mark reported an issue where a thread registers an FAM atomic
region, and the registration does not occur for other threads
in the same process. In order to address this, we'll need to
make the data structure which stores the information about the
registration global instead of on a per-thread basis.

Since the data structure can now be shared between multiple
threads, we'll need to synchronize access to it. This can be
done with a reader/writer lock, to allow multiple threads to
simultaneously look up the list before invoking an atomic
operation. Priority is given to writers, because the read lock can
be taken significantly more often and we must make sure writers
don't get starved when trying to register or unregister a region.

Signed-off-by: Jason Low <jason.low2@hpe.com>