Add new flag for determining if zbridge atomics should be used
The library tracks each register region in a "node" structure.
We'll add an additional boolean flag 'use_zbridge_atomics' to
the node struct which signifies whether to use zbridge atomics
or the simulated atomics for atomics within the registered region.
The value of the flag is determined by whether or not we are able
to perform a zbridge atomic read at the start of the region.

Signed-off-by: Jason Low <jason.low2@hpe.com>