Use simulation if no /dev/fam_atomic device file is found
Use simulated/native atomics instead of exiting with an error if no
/dev/fam_atomic file is found. We'll also remove the #ifdef TMAS. This
reduces complexity in that we don't have to carry both a TMAS and a
non-TMAS package on ARM64. On non-TMAS systems, the device file would
not be found, in which case, it will revert to using simulated atomics.

TMAS users must make sure that the fam atomic driver has been loaded
and that the /dev/fam_atomic device file exists in order for the
library to use zbridge atomics.

Signed-off-by: Jason Low <jason.low2@hpe.com>