tests: Make 'start' a native atomic variable
In fam_atomic_single_node_test, the 'start' atomic variable is
used for managing the test, and is not meant to directly test
the fam atomics. If there is a bug with the fam atomics, the
atomics test would not start, and the test would freeze rather
than report an error. Thus, 'start' will be changed from an
fam atomic to a native atomic.

Signed-off-by: Jason Low <jason.low2@hpe.com>