Add additional 'lfs_fd' field to the ioctl args structure
The ioctl() is implemented in the fam_atomic_driver and not
in LFS. So the fd that gets used in the call to the ioctl()
is the fd associated with the fam_atomic_driver device file,
and not the LFS fd. Thus, we need to also pass the LFS fd
in the ioctl() call. We can do this by adding an additional
'lfs_fd' field to the ioctl args structure.

Signed-off-by: Jason Low <jason.low2@hpe.com>