fam atomics: API modifications
Based on feedback from users of libfam-atomic, the padded atomics
are not being used, yet the unpadded atomic API names are too long.
This patch removes all of the APIs for the "padded" atomics and
renames all of the "unpadded" atomic APIs to not include the term
"unpadded". The result of this is that the library will only
support one set of APIs and that the library never manages the
cache for the user.

The drawback is that this requires all applications will have to
manually make sure atomics don't share cachelines with regular data.
However, this is fine since essentially all users of the library
(at least for now) are using this option.

Signed-off-by: Jason Low <jason.low2@hpe.com>