Check if LFS is really running before using zbridge atomics
The atomics code now queries the tm-fuse module to convert an
LFS file + offset to LZA. If an application uses atomics with
an "/lfs" file without LFS running, the file would really be
an EXT4 file (if EXT4 is the root file system). This results
in a kernel panic when the tm-fuse module tries to make the
LZA conversion with the EXT4 file.

Eventually, lfs_obtain_lza_and_book_offset() will have the
checks for if the 'struct file' corresponds to an LFS file.
In the meantime, we'll add a temporary check in the userspace
library for whether LFS is running when the user tries to use
zbridge atomics in order to avoid kernel panics.

Signed-off-by: Jason Low <jason.low2@hpe.com>