Require non-FAM regions to be registered
Currently, we assume non registered regions to be DRAM data and use
native atomics. However, the only benefit with this is that applications
can avoid registering a region when using fam atomics on DRAM.

Registration isn't expensive, especially since it does not called often. On
the other hand, allowing the library to be used without needing to register
regions can be extremely error prone, and is not worth the trade-off.

This patch modifies the library such that we require users to also
register DRAM files before using atomics on a mapped DRAM file.

Signed-off-by: Jason Low <jason.low2@hpe.com>