doc: Fix incorrect fam lock initialization
In the ERS fam lock example, we initialize the spinlock with
FAM_SPINLOCK_INITIALIZER but it is missing a necessary flush
operation.

Update the example program to use fam_spin_lock_init() which
initializes the lock using an fam atomic operation. The
alternative would be to perform the flush operations, although
init functions are the more common way to initialize non-static
locks.

Also update the documention to mention fam_spin_lock_init()
and that the use of FAM_SPINLOCK_INITILIZER should be followed
by a flush operation.

Signed-off-by: Jason Low <jason.low2@hpe.com>